Add more test for the authentication controller
